병원 정보 관련 python 파일 위치 변경
파일 관리를 위해 폴더 위치 python 폴더 밑으로 변경.
